Update from original

+++ front/ticket.form.php
@@ -1,44 +1,44 @@
 <?php
-/*
- * @version $Id$
- -------------------------------------------------------------------------
- GLPI - Gestionnaire Libre de Parc Informatique
- Copyright (C) 2015 Teclib'.
-
- http://glpi-project.org
-
- based on GLPI - Gestionnaire Libre de Parc Informatique
- Copyright (C) 2003-2014 by the INDEPNET Development Team.
- 
- -------------------------------------------------------------------------
-
- LICENSE
-
- This file is part of GLPI.
-
- GLPI is free software; you can redistribute it and/or modify
- it under the terms of the GNU General Public License as published by
- the Free Software Foundation; either version 2 of the License, or
- (at your option) any later version.
-
- GLPI is distributed in the hope that it will be useful,
- but WITHOUT ANY WARRANTY; without even the implied warranty of
- MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- GNU General Public License for more details.
-
- You should have received a copy of the GNU General Public License
- along with GLPI. If not, see <http://www.gnu.org/licenses/>.
- --------------------------------------------------------------------------
+/**
+ * ---------------------------------------------------------------------
+ * GLPI - Gestionnaire Libre de Parc Informatique
+ * Copyright (C) 2015-2017 Teclib' and contributors.
+ *
+ * http://glpi-project.org
+ *
+ * based on GLPI - Gestionnaire Libre de Parc Informatique
+ * Copyright (C) 2003-2014 by the INDEPNET Development Team.
+ *
+ * ---------------------------------------------------------------------
+ *
+ * LICENSE
+ *
+ * This file is part of GLPI.
+ *
+ * GLPI is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * GLPI is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with GLPI. If not, see <http://www.gnu.org/licenses/>.
+ * ---------------------------------------------------------------------
  */
 
 /** @file
 * @brief
 */
 
+use Glpi\Event;
+
 include ('../inc/includes.php');
 
 Session::checkLoginUser();
-$fup   = new TicketFollowup();
 $track = new Ticket();
 
 if (!isset($_GET['id'])) {
@@ -48,13 +48,6 @@ if (!isset($_GET['id'])) {
 if (isset($_POST["add"])) {
    $track->check(-1, CREATE, $_POST);
 
-   if (isset($_POST["my_items"]) && !empty($_POST["my_items"])) {
-      $splitter = explode("_",$_POST["my_items"]);
-      if (count($splitter) == 2) {
-         $_POST["itemtype"] = $splitter[0];
-         $_POST["items_id"] = $splitter[1];
-      }
-   }
    if ($id = $track->add($_POST)) {
       if ($_SESSION['glpibackcreated']) {
          Html::redirect($track->getFormURL()."?id=".$id);
@@ -65,8 +58,25 @@ if (isset($_POST["add"])) {
 } else if (isset($_POST['update'])) {
    $track->check($_POST['id'], UPDATE);
 
-
    $track->update($_POST);
+
+   if (isset($_POST['kb_linked_id'])) {
+      //if solution should be linked to selected KB entry
+      $params = [
+         'knowbaseitems_id' => $_POST['kb_linked_id'],
+         'itemtype'         => $track->getType(),
+         'items_id'         => $track->getID()
+      ];
+      $existing = $DB->request(
+         'glpi_knowbaseitems_items',
+         $params
+      );
+      if ($existing->numrows() == 0) {
+         $kb_item_item = new KnowbaseItem_Item();
+         $kb_item_item->add($params);
+      }
+   }
+
    Event::log($_POST["id"], "ticket", 4, "tracking",
               //TRANS: %s is the user login
               sprintf(__('%s updates an item'), $_SESSION["glpiname"]));
@@ -112,10 +122,10 @@ if (isset($_POST["add"])) {
    }
    $track->redirectToList();
 
-} else if (isset($_POST['sla_delete'])) {
+} else if (isset($_POST['slt_delete'])) {
    $track->check($_POST["id"], UPDATE);
 
-   $track->deleteSLA($_POST["id"], $_POST['delete_due_date']);
+   $track->deleteSLT($_POST["id"], $_POST['type'], $_POST['delete_date']);
    Event::log($_POST["id"], "ticket", 4, "tracking",
               //TRANS: %s is the user login
               sprintf(__('%s updates an item'), $_SESSION["glpiname"]));
@@ -150,14 +160,17 @@ if (isset($_POST["add"])) {
               sprintf(__('%s adds an actor'), $_SESSION["glpiname"]));
    Html::redirect($CFG_GLPI["root_doc"]."/front/ticket.form.php?id=".$_POST['tickets_id']);
 } else if (isset($_REQUEST['delete_document'])) {
-   $document_item = new Document_Item;
-   $found_document_items = $document_item->find("itemtype = 'Ticket' ".
-                                                " AND items_id = ".intval($_REQUEST['tickets_id']).
-                                                " AND documents_id = ".intval($_REQUEST['documents_id']));
-   foreach ($found_document_items  as $item) {
-      $document_item->delete($item, true);
+   $doc = new Document();
+   $doc->getFromDB(intval($_REQUEST['documents_id']));
+   if ($doc->can($doc->getID(), UPDATE)) {
+      $document_item = new Document_Item;
+      $found_document_items = $document_item->find("itemtype = 'Ticket' ".
+                                                   " AND items_id = ".intval($_REQUEST['tickets_id']).
+                                                   " AND documents_id = ".$doc->getID());
+      foreach ($found_document_items  as $item) {
+         $document_item->delete($item, true);
+      }
    }
-
    Html::back();
 }
 
@@ -191,8 +204,19 @@ if (isset($_GET["id"]) && ($_GET["id"] > 0)) {
    }
 
 } else {
-   Html::header(__('New ticket'),'',"helpdesk","ticket");
+   Html::header(__('New ticket'), '', "helpdesk", "ticket");
    unset($_REQUEST['id']);
+   // alternative email must be empty for create ticket
+   unset($_REQUEST['_users_id_requester_notif']['alternative_email']);
+   unset($_REQUEST['_users_id_observer_notif']['alternative_email']);
+   unset($_REQUEST['_users_id_assign_notif']['alternative_email']);
+   unset($_REQUEST['_suppliers_id_assign_notif']['alternative_email']);
+   // Add a ticket from item : format data
+   if (isset($_REQUEST['_add_fromitem'])
+       && isset($_REQUEST['itemtype'])
+       && isset($_REQUEST['items_id'])) {
+      $_REQUEST['items_id'] = array($_REQUEST['itemtype'] => array($_REQUEST['items_id']));
+   }
    $track->display($_REQUEST);
 }
 
@@ -202,4 +226,3 @@ if ($_SESSION["glpiactiveprofile"]["interface"] == "helpdesk") {
 } else {
    Html::footer();
 }
-?>
